bugfix to magic test

+++ test/magic__test.cpp
@@ -13,7 +13,7 @@ void test_bishop_moves() {
 	position pos;
 	set_pos(&pos,"8/1r3k2/8/3bB3/8/8/8/3K4 b - - 0 1");
 
-	uint bishop_moves = get_bishop_moves(&pos,E5,pos.black_pieces);
+	uint64 bishop_moves = get_bishop_moves(&pos,E5,pos.black_pieces);
 	assert(bishop_moves==0);
 
 	bishop_moves = get_bishop_moves(&pos,D5,pos.black_pieces);
@@ -24,7 +24,7 @@ void test_rook_moves() {
 	position pos;
 	set_pos(&pos,"8/1r3k2/8/3bB3/8/8/8/3K4 b - - 0 1");
 
-	uint rook_moves = get_rook_moves(&pos,B7,pos.black_pieces);
+	uint64 rook_moves = get_rook_moves(&pos,B7,pos.black_pieces);
 	assert(rook_moves==bb_squares[F7]);
 }
